file manager restyle
