+++ client/src/FileManager.module.css
@@ -1,36 +1,56 @@
 .FileManagerWrapper {
     width: 100%;
-    height: calc( 100vh - 108px - 60px - 24px - 56px - 2px - 10px );
+    height: calc(100vh - 108px - 60px - 24px - 56px - 2px - 10px);
 }
+
 .FileManager {
     width: 100%;
     height: 100%;
     border: none;
 }
 
-.ButtonGroup :first-child {
-    border-left-color: transparent;
+.FileToolBar {
+    height: 87px;
+    background-color: #e5e5e5;
 }
-.ButtonGroup :last-child {
-    border-right-color: transparent;
+
+.ButtonGroupWrapper {
+    padding-left: 24px;
+    padding-right: 24px;
+    height: 65px;
+    gap: 12px;
 }
-.ButtonGroup > *, .ButtonGroup > *:disabled {
-    border-left: 1px solid var(--gray);
-    border-right: 1px solid var(--gray);
+
+.ButtonGroupWrapper:not(:first-of-type) {
+    border-left: 1px solid #BEBEBE;
 }
-.ButtonGroup > *:disabled {
+
+.ButtonGroupTitle {
+    margin: 0;
+    color: #3d3d3d;
+    line-height: 16px;
+    font-size: 16px;
+}
+
+.ButtonGroup {
+    gap: 5px;
+}
+
+.ButtonGroup>*:disabled {
     background-color: var(--gray);
-    border-color: var(--gray);
     opacity: .5;
 }
+
 .iconRename i::before {
-    font-family : 'Fontello';
-    content     : '\e802  ';
+    font-family: 'Fontello';
+    content: '\e802  ';
 
     font-style: normal;
     display: inline-block;
     width: 14px;
+    line-height: 14px;
 }
+
 .DropdownItem .fa {
     width: 1rem;
     text-align: center;
