+++ client/src/FileManagerLayout.js
@@ -1,28 +1,70 @@
-import { FullscreenButton } from 'flight-webapp-components';
+import { useContext } from 'react';
+import classNames from 'classnames';
+
+import {
+  DefaultErrorMessage,
+  ErrorBoundary,
+  Overlay,
+  OverlayContainer,
+  Spinner,
+  FullscreenButton
+} from 'flight-webapp-components';
+
+import {
+  Context as FileManagerContext,
+} from './FileManagerContext';
+
+import FileToolbar from './FIleToolbar';
+
+function FileManagerLayout({
+  children
+}) {
+  const { state, currentAbsDir } = useContext(FileManagerContext);
+  let loadingMessage = null;
+  if (state !== 'connected' && state !== 'failed') {
+    loadingMessage = <Loading text="Loading file manager..." />;
+  }
+  let errorMessage = null;
+  if (state === 'failed') {
+    errorMessage = <DefaultErrorMessage />;
+  }
+
+  return (
+      <div className="overflow-auto">
+        <div className="row no-gutters">
+          <div className="col">
+            <div className="card-header toolbar text-light">
+              <div className="row no-gutters">
+                <div className="col">
+                  <div className="d-flex align-items-center">
+                    <h5 className="flex-grow-1 mb-0">
+                      {currentAbsDir}
+                    </h5>
+                    <Toolbar
+                      fileManagerState={state}
+                      onZenChange={ () => {} }
+                    />
+                  </div>
+                </div>
+              </div>
+            </div>
+            <FileToolbar />
+            <div className="bg-white">
+              {loadingMessage}
+              {errorMessage}
+              {children}
+            </div>
+          </div>
+        </div>
+      </div>
+  );
+}
 
 function Toolbar({
-    onDisconnect,
     onFullscreenChange,
-    onReconnect,
     onZenChange,
-    fileManagerState: fileManagerState,
+    fileManagerState
 }) {
-  const disconnectBtn = fileManagerState === 'connected' ? (
-    <i
-      className="fa fa-times ml-2 link white-text"
-      title="Disconnect"
-      onClick={onDisconnect}
-    ></i>
-  ) : null;
-
-  const reconnectBtn = fileManagerState === 'disconnected' ? (
-    <i
-      className="fa fa-bolt ml-2 link white-text"
-      title="Reconnect"
-      onClick={onReconnect}
-    ></i>
-  ) : null;
-
   const fullscreenBtn = fileManagerState === 'connected' ?
     <FullscreenButton
       onFullscreenChange={onFullscreenChange}
@@ -33,50 +75,18 @@ function Toolbar({
   return (
     <div className="btn-toolbar">
       {fullscreenBtn}
-      {disconnectBtn}
-      {reconnectBtn}
     </div>
   );
 }
-  
-function FileManagerLayout({
-    children,
-    onDisconnect,
-    onFullscreenChange,
-    onReconnect,
-    onZenChange,
-    terminalState: fileManagerState,
-    title,
-  }) {
-    return (
-      <div className="overflow-auto">
-        <div className="row no-gutters">
-          <div className="col">
-            <div className="card-header toolbar text-light">
-              <div className="row no-gutters">
-                <div className="col">
-                  <div className="d-flex align-items-center">
-                    <h5 className="flex-grow-1 mb-0">
-                      {title}
-                    </h5>
-                    <Toolbar
-                      fileManagerState={fileManagerState}
-                      onDisconnect={onDisconnect}
-                      onFullscreenChange={onFullscreenChange}
-                      onReconnect={onReconnect}
-                      onZenChange={onZenChange}
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className="bg-black">
-              {children}
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
 
-  export default FileManagerLayout;
+function Loading({ text }) {
+  return (
+    <OverlayContainer>
+      <Overlay>
+        <Spinner text={text} />
+      </Overlay>
+    </OverlayContainer>
+  );
+}
+
+export default FileManagerLayout;
